Enhance line height support
The line height option now accepts a unitless value which is multiplied to the font size (default to `1.2`), a percentage of the font size or a length (e.g. `'12px'` or `1.2em`). Also move the `lineHeight` option under `font.lineHeight` since this value is supposed to be part of the font string (https://developer.mozilla.org/en-US/docs/Web/CSS/font).