Allow labels to react to interactions with elements
The scriptable option context now includes the `active` state of the associated element, which allows to dynamically change the appearance and/or content of the labels when the user interacts with the chart elements. When an element become active (or inactive), only its associated label options are re-evaluated, then the chart re-rendered (no whole `chart.update()` involved).