Implement label events (enter, leave and click) (#40)
Add a new `listeners` option (object), where the property is the type of the event to listen and the value a callback with a unique `context` argument containing the same information as for scriptable options. If a callback explicitly returns `true`, the label is updated with the modified context and the chart re-rendered. This allows to implement visual interactions with labels such as highlight, selection, etc.

Note that there should be no performance penalty for configuration that don't use events since if there is no listener registered, we immediately ignore the incoming event and avoid useless extra computation.