Make the plugin implementation explicitly private
We don't want to be locked in our implementation detail because developers start to write code relying on the expando property. That's why all properties have been made private (prefixed by `_`) and a note has been added to the README to explain that properties / methods prefixed by `_` or `$` are private and should not be accessed.