Prevent overlapping labels when `display: 'auto'`
When the `display` option is `'auto'`, labels are automatically hidden when they overlap (see the docs for the priority rules). The overlap detection uses the Separating Axis Theorem (simplified), implemented in the HitBox class which is now able to intersect with another box.

Move the positioning / overlapping logic in a new Layout singleton (`layout.js`), which is also responsible of drawing all labels in the right order. It also provides a way to lookup for labels at a specific position (`{x,y}`).