Fix hover interaction animations (#119)
Flag the render request as "lazy" so the controller can cancel it in order to correctly animate hover interactions. The 1ms duration is a workaround to make sure an animation is created so the controller can stop it before any transition. Also, verify if there is an active animation, in which case we don't need to explicitly call render().