Move samples in docs and make them editable
Samples are now integrated in the docs, which provides a better user
experience and allows to link samples and guides together. Implement
a Vuepress plugin that displays markdown `js chart-editor` code block
in a Chart editor, in which the user can preview the chart, trigger
actions and edit the source code which is exposed in named "blocks".