Migrate to GitHub Actions
Better release workflow which doesn't rely anymore on a release branch
but instead must be manually triggered from the GH Actions panel for a
specific branch. A tag is automatically created for the version in the
"package.json" file and will be pushed with dist files. Assets (*.zip,
*.tgz, dist/*.js) are attached to the release notes, which must exist
in a "draft" state for the release workflow to success.

Publishing to npm now happens when a GH release is published, which is
not anymore automatic when the Git tag is created. This allows a final
review of the release notes, tag and assets. When publishing to npm, a
tag is calculated based on existing npm tags and can be either "next",
"dev", "latest" or "untagged". See "scripts/utils.js" for details.

Finally, rewrite all scripts to JavaScript (no more bash scripts).